renames the CustomCollapsible component to NestableCollapsible
